+++ lib/slide_container.dart
@@ -1,16 +1,25 @@
 import 'package:flutter/scheduler.dart';
 import 'package:flutter/widgets.dart';
 
-enum VerticalSlideContainerDirection { topToBottom, bottomToTop, bidirectional }
+enum SlideContainerDirection {
+  topToBottom,
+  bottomToTop,
+  vertical,
+  leftToRight,
+  rightToLeft,
+  horizontal
+}
 
-/// Container that can be slid vertically.
+/// Container that can be slid vertically or horizontally.
 ///
 /// Applies a dampening effect to the movement for a smoother gesture.
-class VerticalSlideContainer extends StatefulWidget {
+/// Can validate slide with either drag gesture velocity or distance.
+/// Will automatically finish the slide animation when the drag gesture ends.
+class SlideContainer extends StatefulWidget {
   final Widget child;
 
   /// Constrain the direction the container can be slid.
-  final VerticalSlideContainerDirection slideDirection;
+  final SlideContainerDirection slideDirection;
 
   /// When the gesture ends the animation play at this speed to move back to the start position or to [maxSlideDistance] (see [minSlideDistanceToValidate]).
   /// This takes into account the position of the container just before the gesture ends.
@@ -20,7 +29,7 @@ class VerticalSlideContainer extends StatefulWidget {
 
   /// If not null the container will not slide beyond this value.
   ///
-  /// Default to [MediaQueryData.size].height
+  /// Default to [MediaQueryData.size].height if the [slideDirection] is vertical and [MediaQueryData.size].width if the [slideDirection] is horizontal.
   ///
   /// In px.
   final double maxSlideDistance;
@@ -39,10 +48,11 @@ class VerticalSlideContainer extends StatefulWidget {
   final double minSlideDistanceToValidate;
 
   /// The strength of the dampening effect when the container is moved.
+  ///
   /// The bigger this value the slower the container will move toward the finger position.
+  /// A Value of 1.0 means no damping is added.
   ///
   /// Needs to be superior or equal to 1.0.
-  ///
   final double dampeningStrength;
 
   /// Called when the slide gesture starts.
@@ -55,12 +65,13 @@ class VerticalSlideContainer extends StatefulWidget {
   final VoidCallback onSlideCanceled;
 
   /// Called each frame when the slide gesture is active (i.e. after [onSlideStarted] and before [onSlideCompleted] or [onSlideCanceled]) and during the auto-slide.
-  /// returns the position of the slide container as a value between 0.0 and 1.0 where 0.0 means the container is at the starting position and 1.0 means the container is at [maxSlideDistance].
+  ///
+  /// returns the normalized position of the slide container as a value between 0.0 and 1.0 where 0.0 means the container is at the starting position and 1.0 means the container is at [maxSlideDistance].
   final ValueChanged<double> onSlide;
 
-  VerticalSlideContainer({
+  SlideContainer({
     @required this.child,
-    this.slideDirection = VerticalSlideContainerDirection.bidirectional,
+    this.slideDirection = SlideContainerDirection.vertical,
     this.minDragVelocityForAutoSlide = 600.0,
     this.autoSlideDuration = const Duration(milliseconds: 300),
     this.dampeningStrength = 8.0,
@@ -80,30 +91,42 @@ class VerticalSlideContainer extends StatefulWidget {
   _State createState() => _State();
 }
 
-class _State extends State<VerticalSlideContainer> with TickerProviderStateMixin {
+class _State extends State<SlideContainer> with TickerProviderStateMixin {
   double dragValue = 0.0;
   double dragTarget = 0.0;
   bool isFirstDragFrame;
   AnimationController animationController;
   Ticker followFingerTicker;
 
-  double get maxDragDistance => widget.maxSlideDistance ?? MediaQuery.of(context).size.height;
+  bool get isVerticalSlide =>
+      widget.slideDirection == SlideContainerDirection.topToBottom ||
+      widget.slideDirection == SlideContainerDirection.bottomToTop ||
+      widget.slideDirection == SlideContainerDirection.vertical;
 
-  double get minDragDistanceToValidate => widget.minSlideDistanceToValidate ?? maxDragDistance * 0.5;
+  double get maxDragDistance =>
+      widget.maxSlideDistance ??
+      (isVerticalSlide
+          ? MediaQuery.of(context).size.height
+          : MediaQuery.of(context).size.width);
+
+  double get minDragDistanceToValidate =>
+      widget.minSlideDistanceToValidate ?? maxDragDistance * 0.5;
 
   @override
   void initState() {
-    animationController = AnimationController(vsync: this, duration: widget.autoSlideDuration)
-      ..addListener(() {
-        if (widget.onSlide != null) widget.onSlide(animationController.value);
-        setState(() {});
-      });
+    animationController =
+        AnimationController(vsync: this, duration: widget.autoSlideDuration)
+          ..addListener(() {
+            if (widget.onSlide != null)
+              widget.onSlide(animationController.value);
+            setState(() {});
+          });
 
     followFingerTicker = createTicker((_) {
       if ((dragValue - dragTarget).abs() <= 1.0) {
         dragTarget = dragValue;
       } else {
-        /// This dampen the drag movement (acts like a spring, the farther from the finger position the faster it moves toward it).
+        // This dampen the drag movement (acts like a spring, the farther from the finger position the faster it moves toward it).
         dragTarget += (dragValue - dragTarget) / widget.dampeningStrength;
       }
       animationController.value = dragTarget.abs() / maxDragDistance;
@@ -117,6 +140,13 @@ class _State extends State<VerticalSlideContainer> with TickerProviderStateMixin
     followFingerTicker?.dispose();
     super.dispose();
   }
+
+  double getVelocity(DragEndDetails details) => isVerticalSlide
+      ? details.velocity.pixelsPerSecond.dy
+      : details.velocity.pixelsPerSecond.dx;
+
+  double getDelta(DragUpdateDetails details) =>
+      isVerticalSlide ? details.delta.dy : details.delta.dx;
 
   void completeSlide() => animationController.forward().then((_) {
         if (widget.onSlideCompleted != null) widget.onSlideCompleted();
@@ -141,21 +171,28 @@ class _State extends State<VerticalSlideContainer> with TickerProviderStateMixin
       return;
     }
 
-    dragValue = (dragValue + details.delta.dy).clamp(-maxDragDistance, maxDragDistance);
-    if (widget.slideDirection == VerticalSlideContainerDirection.topToBottom) {
+    dragValue = (dragValue + getDelta(details))
+        .clamp(-maxDragDistance, maxDragDistance);
+    if (widget.slideDirection == SlideContainerDirection.topToBottom ||
+        widget.slideDirection == SlideContainerDirection.leftToRight) {
       dragValue = dragValue.clamp(0.0, maxDragDistance);
-    } else if (widget.slideDirection == VerticalSlideContainerDirection.bottomToTop) {
+    } else if (widget.slideDirection == SlideContainerDirection.bottomToTop ||
+        widget.slideDirection == SlideContainerDirection.rightToLeft) {
       dragValue = dragValue.clamp(-maxDragDistance, 0.0);
     }
   }
 
   void handlePanEnd(DragEndDetails details) {
-    if (details.velocity.pixelsPerSecond.dy * dragTarget.sign > widget.minDragVelocityForAutoSlide) {
+    if (getVelocity(details) * dragTarget.sign >
+        widget.minDragVelocityForAutoSlide) {
       completeSlide();
-    } else if (details.velocity.pixelsPerSecond.dy * dragTarget.sign < -widget.minDragVelocityForAutoSlide) {
+    } else if (getVelocity(details) * dragTarget.sign <
+        -widget.minDragVelocityForAutoSlide) {
       cancelSlide();
     } else {
-      dragTarget.abs() > minDragDistanceToValidate ? completeSlide() : cancelSlide();
+      dragTarget.abs() > minDragDistanceToValidate
+          ? completeSlide()
+          : cancelSlide();
     }
     followFingerTicker.stop();
   }
@@ -166,7 +203,15 @@ class _State extends State<VerticalSlideContainer> with TickerProviderStateMixin
         onPanUpdate: handlePanUpdate,
         onPanEnd: handlePanEnd,
         child: Transform.translate(
-          offset: Offset(0.0, animationController.value * maxDragDistance * dragTarget.sign),
+          offset: isVerticalSlide
+              ? Offset(
+                  0.0,
+                  animationController.value * maxDragDistance * dragTarget.sign,
+                )
+              : Offset(
+                  animationController.value * maxDragDistance * dragTarget.sign,
+                  0.0,
+                ),
           child: widget.child,
         ),
       );