+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.1 - 2018-10-06
+
+* Added support for horizontal sliding.
+* **Breaking change**
+    * Renamed class `VerticalSlideContainer` to `SlideContainer`
+    * Renamed enum `VerticalSlideContainerDirection` to `SlideContainerDirection` and changed its values.
+
 ## 1.0.0 - 2018-10-05
 
 * Initial release.